+++ prompt.go
@@ -37,12 +37,15 @@ type ExecuteOnEnterCallback func(input string, indentSize int) (indent int, exec
 
 // Completer is a function that returns
 // a slice of suggestions for the given Document.
-type Completer func(Document) []Suggest
+//
+// startChar and endChar represent the indices of the first and last rune of the text
+// that the suggestions were generated for and that should be replaced by the selected suggestion.
+type Completer func(Document) (suggestions []Suggest, startChar, endChar istrings.RuneNumber)
 
 // Prompt is a core struct of go-prompt.
 type Prompt struct {
 	reader                 Reader
-	buf                    *Buffer
+	Buffer                 *Buffer
 	renderer               *Renderer
 	executor               Executor
 	history                *History
@@ -55,6 +58,7 @@ type Prompt struct {
 	exitChecker            ExitChecker
 	executeOnEnterCallback ExecuteOnEnterCallback
 	skipClose              bool
+	completionReset        bool
 }
 
 // UserInput is the struct that contains the user input context.
@@ -71,10 +75,10 @@ func (p *Prompt) Run() {
 	defer p.Close()
 
 	if p.completion.showAtStart {
-		p.completion.Update(*p.buf.Document())
+		p.completion.Update(*p.Buffer.Document())
 	}
 
-	p.renderer.Render(p.buf, p.completion, p.lexer)
+	p.renderer.Render(p.Buffer, p.completion, p.lexer)
 
 	bufCh := make(chan []byte, 128)
 	stopReadBufCh := make(chan struct{})
@@ -88,12 +92,12 @@ func (p *Prompt) Run() {
 	for {
 		select {
 		case b := <-bufCh:
-			if shouldExit, e := p.feed(b); shouldExit {
-				p.renderer.BreakLine(p.buf, p.lexer)
+			if shouldExit, rerender, input := p.feed(b); shouldExit {
+				p.renderer.BreakLine(p.Buffer, p.lexer)
 				stopReadBufCh <- struct{}{}
 				stopHandleSignalCh <- struct{}{}
 				return
-			} else if e != nil {
+			} else if input != nil {
 				// Stop goroutine to run readBuffer function
 				stopReadBufCh <- struct{}{}
 				stopHandleSignalCh <- struct{}{}
@@ -101,13 +105,13 @@ func (p *Prompt) Run() {
 				// Unset raw mode
 				// Reset to Blocking mode because returned EAGAIN when still set non-blocking mode.
 				debug.AssertNoError(p.reader.Close())
-				p.executor(e.input)
+				p.executor(input.input)
 
-				p.completion.Update(*p.buf.Document())
+				p.completion.Update(*p.Buffer.Document())
 
-				p.renderer.Render(p.buf, p.completion, p.lexer)
+				p.renderer.Render(p.Buffer, p.completion, p.lexer)
 
-				if p.exitChecker != nil && p.exitChecker(e.input, true) {
+				if p.exitChecker != nil && p.exitChecker(input.input, true) {
 					p.skipClose = true
 					return
 				}
@@ -115,15 +119,19 @@ func (p *Prompt) Run() {
 				debug.AssertNoError(p.reader.Open())
 				go p.readBuffer(bufCh, stopReadBufCh)
 				go p.handleSignals(exitCh, winSizeCh, stopHandleSignalCh)
-			} else {
-				p.completion.Update(*p.buf.Document())
-				p.renderer.Render(p.buf, p.completion, p.lexer)
+			} else if rerender {
+				if p.completion.shouldUpdate {
+					p.completion.Update(*p.Buffer.Document())
+				}
+				p.renderer.Render(p.Buffer, p.completion, p.lexer)
 			}
 		case w := <-winSizeCh:
 			p.renderer.UpdateWinSize(w)
-			p.renderer.Render(p.buf, p.completion, p.lexer)
+			p.Buffer.ResetStartLine()
+			p.Buffer.RecalculateStartLine(p.renderer.UserInputColumns(), int(p.renderer.row))
+			p.renderer.Render(p.Buffer, p.completion, p.lexer)
 		case code := <-exitCh:
-			p.renderer.BreakLine(p.buf, p.lexer)
+			p.renderer.BreakLine(p.Buffer, p.lexer)
 			p.Close()
 			os.Exit(code)
 		default:
@@ -138,109 +146,135 @@ func Log(format string, a ...any) {
 		log.Fatalf("error opening file: %v", err)
 	}
 	defer f.Close()
-	fmt.Fprintf(f, format, a...)
+	fmt.Fprintf(f, format+"\n", a...)
 }
 
-func (p *Prompt) feed(b []byte) (shouldExit bool, userInput *UserInput) {
+func (p *Prompt) feed(b []byte) (shouldExit bool, rerender bool, userInput *UserInput) {
 	key := GetKey(b)
-	p.buf.lastKeyStroke = key
+	p.Buffer.lastKeyStroke = key
 	// completion
 	completing := p.completion.Completing()
-	p.handleCompletionKeyBinding(b, key, completing)
+	if p.handleCompletionKeyBinding(b, key, completing) {
+		return false, true, nil
+	}
+
+	cols := p.renderer.UserInputColumns()
+	rows := p.renderer.row
 
 	switch key {
 	case Enter, ControlJ, ControlM:
-		indent, execute := p.executeOnEnterCallback(p.buf.Text(), p.renderer.indentSize)
+		indent, execute := p.executeOnEnterCallback(p.Buffer.Text(), p.renderer.indentSize)
 		if !execute {
-			p.buf.NewLine(false)
+			p.Buffer.NewLine(cols, rows, false)
+
 			var indentStrBuilder strings.Builder
 			indentUnitCount := indent * p.renderer.indentSize
 			for i := 0; i < indentUnitCount; i++ {
 				indentStrBuilder.WriteRune(IndentUnit)
 			}
-			p.buf.InsertText(indentStrBuilder.String(), false, true)
+			p.Buffer.InsertTextMoveCursor(indentStrBuilder.String(), cols, rows, false)
 			break
 		}
 
-		p.renderer.BreakLine(p.buf, p.lexer)
-		userInput = &UserInput{input: p.buf.Text()}
-		p.buf = NewBuffer()
+		p.renderer.BreakLine(p.Buffer, p.lexer)
+		userInput = &UserInput{input: p.Buffer.Text()}
+		p.Buffer = NewBuffer()
 		if userInput.input != "" {
 			p.history.Add(userInput.input)
 		}
 	case ControlC:
-		p.renderer.BreakLine(p.buf, p.lexer)
-		p.buf = NewBuffer()
+		p.renderer.BreakLine(p.Buffer, p.lexer)
+		p.Buffer = NewBuffer()
 		p.history.Clear()
 	case Up, ControlP:
-		line := p.buf.Document().CursorPositionRow()
+		line := p.Buffer.Document().CursorPositionRow()
 		if line > 0 {
-			p.buf.CursorUp(1)
-			break
+			rerender = p.CursorUp(1)
+			return false, rerender, nil
 		}
 		if completing {
 			break
 		}
 
-		if newBuf, changed := p.history.Older(p.buf); changed {
-			p.buf = newBuf
+		if newBuf, changed := p.history.Older(p.Buffer, cols, rows); changed {
+			p.Buffer = newBuf
 		}
 
 	case Down, ControlN:
-		endOfTextRow := p.buf.Document().TextEndPositionRow()
-		row := p.buf.Document().CursorPositionRow()
+		endOfTextRow := p.Buffer.Document().TextEndPositionRow()
+		row := p.Buffer.Document().CursorPositionRow()
 		if endOfTextRow > row {
-			p.buf.CursorDown(1)
-			break
+			rerender = p.CursorDown(1)
+			return false, rerender, nil
 		}
 
 		if completing {
 			break
 		}
 
-		if newBuf, changed := p.history.Newer(p.buf); changed {
-			p.buf = newBuf
+		if newBuf, changed := p.history.Newer(p.Buffer, cols, rows); changed {
+			p.Buffer = newBuf
 		}
-		return
+		return false, true, nil
 	case ControlD:
-		if p.buf.Text() == "" {
-			shouldExit = true
-			return
+		if p.Buffer.Text() == "" {
+			return true, true, nil
 		}
 	case NotDefined:
-		if p.handleASCIICodeBinding(b) {
-			return
+		var checked bool
+		checked, rerender = p.handleASCIICodeBinding(b, cols, rows)
+
+		if checked {
+			return false, rerender, nil
 		}
 		char, _ := utf8.DecodeRune(b)
 		if unicode.IsControl(char) {
-			return
+			return false, false, nil
 		}
 
-		p.buf.InsertText(string(b), false, true)
+		p.Buffer.InsertTextMoveCursor(string(b), cols, rows, false)
 	}
 
-	shouldExit = p.handleKeyBinding(key)
-	return
+	shouldExit, rerender = p.handleKeyBinding(key, cols, rows)
+	return shouldExit, rerender, userInput
 }
 
-func (p *Prompt) handleCompletionKeyBinding(b []byte, key Key, completing bool) {
+func (p *Prompt) handleCompletionKeyBinding(b []byte, key Key, completing bool) (handled bool) {
+	p.completion.shouldUpdate = true
+	cols := p.renderer.UserInputColumns()
+	rows := p.renderer.row
+	completionLen := len(p.completion.tmp)
+	p.completionReset = false
+
 keySwitch:
 	switch key {
 	case Down:
 		if completing || p.completionOnDown {
-			p.completion.Next()
+			p.updateSuggestions(func() {
+				p.completion.Next()
+			})
+			return true
 		}
 	case ControlI:
-		p.completion.Next()
+		p.updateSuggestions(func() {
+			p.completion.Next()
+		})
+		return true
 	case Up:
 		if completing {
-			p.completion.Previous()
+			p.updateSuggestions(func() {
+				p.completion.Previous()
+			})
+			return true
 		}
 	case Tab:
-		if len(p.completion.GetSuggestions()) > 0 {
+		if completionLen > 0 {
 			// If there are any suggestions, select the next one
-			p.completion.Next()
-			break
+			p.updateSuggestions(func() {
+				p.completion.Next()
+			})
+
+			return true
 		}
 
 		// if there are no suggestions insert indentation
@@ -255,39 +289,95 @@ keySwitch:
 				newBytes = append(newBytes, byt)
 			}
 		}
-		p.buf.InsertText(string(newBytes), false, true)
+		p.Buffer.InsertTextMoveCursor(string(newBytes), cols, rows, false)
+		return true
 	case BackTab:
-		if len(p.completion.GetSuggestions()) > 0 {
+		if completionLen > 0 {
 			// If there are any suggestions, select the previous one
-			p.completion.Previous()
-			break
+			p.updateSuggestions(func() {
+				p.completion.Previous()
+			})
+			return true
 		}
 
-		text := p.buf.Document().CurrentLineBeforeCursor()
+		text := p.Buffer.Document().CurrentLineBeforeCursor()
 		for _, char := range text {
 			if char != IndentUnit {
 				break keySwitch
 			}
 		}
-		p.buf.DeleteBeforeCursor(istrings.RuneNumber(p.renderer.indentSize))
+		p.Buffer.DeleteBeforeCursor(istrings.RuneNumber(p.renderer.indentSize), cols, rows)
+		return true
 	default:
 		if s, ok := p.completion.GetSelectedSuggestion(); ok {
-			w := p.buf.Document().GetWordBeforeCursorUntilSeparator(p.completion.wordSeparator)
+			w := p.Buffer.Document().GetWordBeforeCursorUntilSeparator(p.completion.wordSeparator)
 			if w != "" {
-				p.buf.DeleteBeforeCursor(istrings.RuneNumber(len([]rune(w))))
+				p.Buffer.DeleteBeforeCursor(istrings.RuneNumber(len([]rune(w))), cols, rows)
 			}
-			p.buf.InsertText(s.Text, false, true)
+			p.Buffer.InsertTextMoveCursor(s.Text, cols, rows, false)
+		}
+		if completionLen > 0 {
+			p.completionReset = true
 		}
 		p.completion.Reset()
 	}
+	return false
 }
 
-func (p *Prompt) handleKeyBinding(key Key) bool {
-	shouldExit := false
+func (p *Prompt) updateSuggestions(fn func()) {
+	cols := p.renderer.UserInputColumns()
+	rows := p.renderer.row
+
+	prevStart := p.completion.startCharIndex
+	prevEnd := p.completion.endCharIndex
+	prevSuggestion, prevSelected := p.completion.GetSelectedSuggestion()
+
+	fn()
+
+	p.completion.shouldUpdate = false
+	newSuggestion, newSelected := p.completion.GetSelectedSuggestion()
+
+	// do nothing
+	if !prevSelected && !newSelected {
+		return
+	}
+
+	// insert the new selection
+	if !prevSelected {
+		p.Buffer.DeleteBeforeCursor(p.completion.endCharIndex-p.completion.startCharIndex, cols, rows)
+		p.Buffer.InsertTextMoveCursor(newSuggestion.Text, cols, rows, false)
+		return
+	}
+	// delete the previous selection
+	if !newSelected {
+		p.Buffer.DeleteBeforeCursor(
+			istrings.RuneCount(prevSuggestion.Text)-(prevEnd-prevStart),
+			cols,
+			rows,
+		)
+		return
+	}
+
+	// delete previous selection and render the new one
+	p.Buffer.DeleteBeforeCursor(
+		istrings.RuneCount(prevSuggestion.Text),
+		cols,
+		rows,
+	)
+
+	p.Buffer.InsertTextMoveCursor(newSuggestion.Text, cols, rows, false)
+}
+
+func (p *Prompt) handleKeyBinding(key Key, cols istrings.Width, rows int) (shouldExit bool, rerender bool) {
+	var executed bool
 	for i := range commonKeyBindings {
 		kb := commonKeyBindings[i]
 		if kb.Key == key {
-			kb.Fn(p.buf)
+			result := kb.Fn(p)
+			executed = true
+			if !rerender {
+				rerender = result
+			}
 		}
 	}
 
@@ -296,7 +386,11 @@ func (p *Prompt) handleKeyBinding(key Key) bool {
 		for i := range emacsKeyBindings {
 			kb := emacsKeyBindings[i]
 			if kb.Key == key {
-				kb.Fn(p.buf)
+				result := kb.Fn(p)
+				executed = true
+				if !rerender {
+					rerender = result
+				}
 			}
 		}
 	}
@@ -305,24 +399,33 @@ func (p *Prompt) handleKeyBinding(key Key) bool {
 	for i := range p.keyBindings {
 		kb := p.keyBindings[i]
 		if kb.Key == key {
-			kb.Fn(p.buf)
+			result := kb.Fn(p)
+			executed = true
+			if !rerender {
+				rerender = result
+			}
 		}
 	}
-	if p.exitChecker != nil && p.exitChecker(p.buf.Text(), false) {
+	if p.exitChecker != nil && p.exitChecker(p.Buffer.Text(), false) {
 		shouldExit = true
 	}
-	return shouldExit
+	if !executed && !rerender {
+		rerender = true
+	}
+	return shouldExit, rerender
 }
 
-func (p *Prompt) handleASCIICodeBinding(b []byte) bool {
-	checked := false
+func (p *Prompt) handleASCIICodeBinding(b []byte, cols istrings.Width, rows int) (checked, rerender bool) {
 	for _, kb := range p.ASCIICodeBindings {
 		if bytes.Equal(kb.ASCIICode, b) {
-			kb.Fn(p.buf)
+			result := kb.Fn(p)
+			if !rerender {
+				rerender = result
+			}
 			checked = true
 		}
 	}
-	return checked
+	return checked, rerender
 }
 
 // Input starts the prompt, lets the user
@@ -334,10 +437,10 @@ func (p *Prompt) Input() string {
 	defer p.Close()
 
 	if p.completion.showAtStart {
-		p.completion.Update(*p.buf.Document())
+		p.completion.Update(*p.Buffer.Document())
 	}
 
-	p.renderer.Render(p.buf, p.completion, p.lexer)
+	p.renderer.Render(p.Buffer, p.completion, p.lexer)
 	bufCh := make(chan []byte, 128)
 	stopReadBufCh := make(chan struct{})
 	go p.readBuffer(bufCh, stopReadBufCh)
@@ -345,17 +448,17 @@ func (p *Prompt) Input() string {
 	for {
 		select {
 		case b := <-bufCh:
-			if shouldExit, e := p.feed(b); shouldExit {
-				p.renderer.BreakLine(p.buf, p.lexer)
+			if shouldExit, rerender, input := p.feed(b); shouldExit {
+				p.renderer.BreakLine(p.Buffer, p.lexer)
 				stopReadBufCh <- struct{}{}
 				return ""
-			} else if e != nil {
+			} else if input != nil {
 				// Stop goroutine to run readBuffer function
 				stopReadBufCh <- struct{}{}
-				return e.input
-			} else {
-				p.completion.Update(*p.buf.Document())
-				p.renderer.Render(p.buf, p.completion, p.lexer)
+				return input.input
+			} else if rerender {
+				p.completion.Update(*p.Buffer.Document())
+				p.renderer.Render(p.Buffer, p.completion, p.lexer)
 			}
 		default:
 			time.Sleep(10 * time.Millisecond)
@@ -414,6 +517,82 @@ func (p *Prompt) setup() {
 	debug.AssertNoError(p.reader.Open())
 	p.renderer.Setup()
 	p.renderer.UpdateWinSize(p.reader.GetWinSize())
+}
+
+// Move to the left on the current line.
+// Returns true when the view should be rerendered.
+func (p *Prompt) CursorLeft(count istrings.RuneNumber) bool {
+	b := p.Buffer
+	cols := p.renderer.UserInputColumns()
+	previousCursor := b.DisplayCursorPosition(cols)
+
+	rerender := p.Buffer.CursorLeft(count, cols, p.renderer.row) || p.completionReset || len(p.completion.tmp) > 0
+	if rerender {
+		return true
+	}
+
+	newCursor := b.DisplayCursorPosition(cols)
+	p.renderer.previousCursor = newCursor
+	p.renderer.move(previousCursor, newCursor)
+	p.renderer.flush()
+	return false
+}
+
+// Move the cursor to the right on the current line.
+// Returns true when the view should be rerendered.
+func (p *Prompt) CursorRight(count istrings.RuneNumber) bool {
+	b := p.Buffer
+	cols := p.renderer.UserInputColumns()
+	previousCursor := b.DisplayCursorPosition(cols)
+
+	rerender := p.Buffer.CursorRight(count, cols, p.renderer.row) || p.completionReset || len(p.completion.tmp) > 0
+	if rerender {
+		return true
+	}
+
+	newCursor := b.DisplayCursorPosition(cols)
+	p.renderer.previousCursor = newCursor
+	p.renderer.move(previousCursor, newCursor)
+	p.renderer.flush()
+	return false
+}
+
+// Move the cursor up.
+// Returns true when the view should be rerendered.
+func (p *Prompt) CursorUp(count int) bool {
+	b := p.Buffer
+	cols := p.renderer.UserInputColumns()
+	previousCursor := b.DisplayCursorPosition(cols)
+
+	rerender := p.Buffer.CursorUp(count, cols, p.renderer.row) || p.completionReset || len(p.completion.tmp) > 0
+	if rerender {
+		return true
+	}
+
+	newCursor := b.DisplayCursorPosition(cols)
+	p.renderer.previousCursor = newCursor
+	p.renderer.move(previousCursor, newCursor)
+	p.renderer.flush()
+	return false
+}
+
+// Move the cursor down.
+// Returns true when the view should be rerendered.
+func (p *Prompt) CursorDown(count int) bool {
+	b := p.Buffer
+	cols := p.renderer.UserInputColumns()
+	previousCursor := b.DisplayCursorPosition(cols)
+
+	rerender := p.Buffer.CursorDown(count, cols, p.renderer.row) || p.completionReset || len(p.completion.tmp) > 0
+	if rerender {
+		return true
+	}
+
+	newCursor := b.DisplayCursorPosition(cols)
+	p.renderer.previousCursor = newCursor
+	p.renderer.move(previousCursor, newCursor)
+	p.renderer.flush()
+	return false
 }
 
 func (p *Prompt) Close() {